requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
av==16.1.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
ctranslate2==4.7.1
filelock==3.25.0
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
onnxruntime==1.24.3
openai-whisper==20250625
opencv-python==4.13.0.92
packaging==26.0
pillow==12.0.0
protobuf==7.34.0
Pygments==2.19.2
pytorch-triton-rocm==3.1.0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
setuptools==82.0.0
shellingham==1.5.4
sympy==1.13.1
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.5.1+rocm6.2
torchaudio==2.5.1+rocm6.2
torchvision==0.20.1+rocm6.2
tqdm==4.67.3
triton==3.6.0
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
